NyaySetu/NyaySetu/NyaySetu/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
PyJWT==2.8.0
passlib==1.7.4
requests==2.32.3
reportlab==4.2.5
deep-translator==1.11.4
google-auth==2.33.0
google-generativeai==0.3.2
